+++ packages/fonts/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@repo/next-fonts",
+  "name": "@repo/fonts",
   "private": true,
   "version": "0.0.0",
   "type": "module",
