+++ packages/mui/package.json
@@ -26,7 +26,7 @@
   },
   "devDependencies": {
     "@repo/config": "workspace:*",
-    "@repo/next-fonts": "workspace:*",
+    "@repo/fonts": "workspace:*",
     "@types/node": "^24.1.0",
     "@types/react": "^19.1.8",
     "@types/react-dom": "^19.1.6",
